refactor: bianca zapatka and eat this use same ingredient parser
