+++ lib/src/recipe_scripts/eat_this.dart
@@ -3,12 +3,12 @@ import 'package:html/dom.dart';
 
 import '../models/ingredient.dart';
 import '../models/ingredient_parsing_result.dart';
-import '../models/meta_data_log.dart';
 import '../models/recipe.dart';
 import '../models/recipe_parsing_job.dart';
 import '../models/recipe_parsing_result.dart';
 import 'parsing_helper.dart';
 import 'recipe_scripts_helper.dart';
+import 'wordpress_ingredient_parsing.dart';
 
 List<String> _notSupportedUrls = [
   "https://www.eat-this.org/wie-man-eine-vegane-kaeseplatte-zusammenstellt/",
@@ -174,12 +174,12 @@ RecipeParsingResult _parseRecipeOldDesign(
 /// If the parsing fails the ingredient in [IngredientParsingResult] will be
 /// null and a suitable log will be returned.
 IngredientParsingResult parseIngredientOldDesign(
-  Element ingredientElement,
+  Element element,
   double servingsMultiplier,
   String recipeUrl, {
   String? language,
 }) {
-  var ingredientText = ingredientElement.text.trim();
+  var ingredientText = element.text.trim();
 
   // Are there two ingredients
   if (ingredientText.contains("+")) {
@@ -299,63 +299,17 @@ RecipeParsingResult _parseRecipeNewDesign(
 /// If the parsing fails the ingredient in [IngredientParsingResult] will be
 /// null and a suitable log will be returned.
 IngredientParsingResult parseIngredientNewDesign(
-  Element ingredientElement,
+  Element element,
   double servingsMultiplier,
   String recipeUrl, {
   String? language,
-}) {
-  var name = "";
-  var nameElements =
-      ingredientElement.getElementsByClassName("wprm-recipe-ingredient-name");
-  if (nameElements.isNotEmpty) {
-    var nameElement = nameElements.first;
-    // Sometimes the name has a url reference in a <a> tag
-    if (nameElement.children.isNotEmpty) {
-      name = nameElement.children.map((element) => element.text).join();
-    } else {
-      name = nameElement.text.trim();
-    }
-  } else {
-    return createFailedIngredientParsingResult(recipeUrl);
-  }
-
-  var logs = <MetaDataLog>[];
-
-  var amount = 0.0;
-  var amountElements =
-      ingredientElement.getElementsByClassName("wprm-recipe-ingredient-amount");
-  if (amountElements.isNotEmpty) {
-    var amountElement = amountElements.first;
-    var amountString = amountElement.text.trim();
-    var parsedAmount = tryParseAmountString(amountString, language: language);
-    if (parsedAmount != null) {
-      amount = parsedAmount * servingsMultiplier;
-    } else {
-      logs.add(
-        createFailedAmountParsingMetaDataLog(
-          recipeUrl,
-          amountString,
-          name,
-        ),
-      );
-    }
-  }
-
-  var unit = "";
-  var unitElements =
-      ingredientElement.getElementsByClassName("wprm-recipe-ingredient-unit");
-  if (unitElements.isNotEmpty) {
-    var unitElement = unitElements.first;
-    unit = unitElement.text.trim();
-  }
-
-  return IngredientParsingResult(
-    ingredients: [
-      Ingredient(amount: amount, unit: unit, name: name),
-    ],
-    metaDataLogs: logs,
-  );
-}
+}) =>
+    parseWordPressIngredient(
+      element,
+      servingsMultiplier,
+      recipeUrl,
+      language: language,
+    );
 
 List<String> _breakUpNumberAndText(String numberAndTextString) {
   var result = <String>[
